refacto adding employees to project without refresh the page, and cleaning code

+++ src/components/Project.js
@@ -1,6 +1,6 @@
 import axios from 'axios'
 import { useEffect, useState } from "react"
-import { useNavigate, useLocation } from "react-router-dom"
+import { useLocation } from "react-router-dom"
 import ProjectEmployeesItem from "./ProjectEmployeesItem"
 import ProjectItem from './ProjectItem'
 import {API_URL} from '../globals'
@@ -10,80 +10,78 @@ function Project() {
     const project = location.state?.project ?? null
     const [employees, updateEmployees] = useState([])
     const [addingEmployees, updateAddingEmployees] = useState([])
-    const [isCheck, updateIsCheck] = useState({})
+    const [elementChecked, updateElementChecked] = useState([])
     const [requireAdding,updateRequireAdding] = useState(false) // if we want to add employees
     useEffect(() => {
-        if(Object.keys(isCheck).length === 0) {
+        if(elementChecked.length === 0 && project) {
             axios.get(API_URL + 'project/' + project.id + "/employees") // getting all employees of this project
                 .then(response => updateEmployees(response.data))
                 .catch(error => console.log(error))
         }
-    }, [isCheck])
+    }, [])
 
     useEffect(() => {
         if(requireAdding) {
             axios.get(API_URL + "employees") // getting all employees
                 .then(response => updateAddingEmployees(response.data))
                 .catch(error => console.log(error))
-            const initialState = {}
-            addingEmployees.forEach((item) => {
-                if(!employees.includes(item)) {
-                    initialState[item.id] = false;
-                }
-            })
-            updateIsCheck(initialState)
         }
-    }, [requireAdding], [])
+    }, [requireAdding])
 
     const handleAddEmployee = () => {
         updateRequireAdding(prevState => !prevState)
     }
 
     const handleCheckboxChange = (id) => {
-        updateIsCheck((prevState) => ({
-        ...prevState,
-        [id]: !prevState[id]
-        }));
+        updateElementChecked((prev) => {
+            if(prev.includes(id)) {
+                return prev.filter((item) => item === id)
+            } else {
+                return [...prev, id]
+            }
+        })
     }
 
     const handleAddingClick = () => {
         updateRequireAdding(prevState => !prevState)
-        Object.keys(isCheck).forEach((key) => {
-            axios.post(API_URL + "employee/" + key + "/project/" + project.id)
-            .then(response => {
-                console.log(response.data)
-                /*if(response.status == 200) {
-                    updateEmployees((prev) => 
-                        prev.push(
-                            addingEmployees.find((item) => item.id === key)
+        if(project) {
+            elementChecked.forEach((key) => {
+                axios.post(API_URL + "employee/" + key + "/project/" + project.id)
+                .then(response => {
+                    console.log(response.data)
+                    if(response.status === 200) {
+                        updateEmployees((prev) => {
+                            return [...prev, addingEmployees.find((item) => item.id == key)]
+                        }
                         )
-                    )
-                }
-                    //TODO : add employee without refreshing the page
-                */
+                    }
+                
+                })
+                .catch(error => console.log(error))
             })
-            .catch(error => console.log(error))
-        })
-        updateIsCheck({})
+        }
+        updateElementChecked([])
     }
 
     const handleDeleteEmployee = (employeeId) => {
-        axios.delete(API_URL + 'employee/' + employeeId + "/project/" + project.id)
-        .then((res) => {
-            if(res.status == 200) {
-                updateEmployees((prev) => 
-                    prev.filter((item) =>
-                        item.id !== employeeId
-                    )
-            )
-            }
-        })
+        if(project) {
+            axios.delete(API_URL + 'employee/' + employeeId + "/project/" + project.id)
+            .then((res) => {
+                if(res.status === 200) {
+                    updateEmployees((prev) => 
+                        prev.filter((item) =>
+                            item.id !== employeeId
+                        )
+                )
+                }
+            })
+        }
     }
 
     function isPresent(id) {
         let isEqual = false
         employees.forEach((item) => {
-            if(item.id == id) {
+            if(item.id === id) {
                 isEqual = true
             }
         })
@@ -104,7 +102,7 @@ function Project() {
                         <label key={data.id} style={{ display: "block" }}>
                         <input
                             type="checkbox"
-                            checked={isCheck[data.id] || false}
+                            checked={elementChecked.includes(data.id)}
                             onChange={() => handleCheckboxChange(data.id)}
                         />
                         <p>{data.firstName + " " + data.lastName}</p>
@@ -116,7 +114,7 @@ function Project() {
             }
             {
                 employees.map((data) => 
-                    <ProjectEmployeesItem key={data.id} employee={data} project={project} onDelete={handleDeleteEmployee}/> // peut etre modifié handleClick dans EmployeeItem
+                    <ProjectEmployeesItem key={data.id} employee={data} project={project} onDelete={handleDeleteEmployee}/>
                 )
             }
         </div>
